Eventorr/Eventorr Project/requirements.txt
Django==3.0.5
django-crispy-forms==1.9.1
twilio==6.44.2
Pillow==7.1.1